+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## WIP `v4.4.1`
+
+### 🐞 Fixes
+
+- fix: [`enforces-shorthand` autofix drops the important modifier when written in `v4` suffix position](https://github.com/francoismassart/eslint-plugin-tailwindcss/issues/476)
+- fix: add a new setting `useLocalPkgWorkaround` to set to false if you want to avoid using `import.meta.url` to paths in TailwindUtils
+
 ## `v4.4.0`
 
 ### 🚀 Features